fix: remove lib document in /api
